Classifier moved to its own folder.
