+++ classifier/src/visualize_classifier_results.py
@@ -6,8 +6,8 @@ import torch
 from pandas import read_csv
 from torchinfo import summary
 
-from src.classifier.classifier import Classifier
-from src.tune_classifier import load_cifar_dataset, prepare_classifier_params
+from classifier.src.classifier.classifier import Classifier
+from classifier.src.tune_classifier import load_cifar_dataset, prepare_classifier_params
 from ml_common.common_utils import AppLog
